+++ common/ChangeLog.md
@@ -2,3 +2,4 @@
 
 * This module now consists of functionalities to support service client library modules.
 * All namespaces in this module now begin with "Azure::Storage::Common" instead of "Azure::Storage".
+* Resolved an issue where user tries to access `Azure::Storage::Common::Default::signer` would throw `undefined method 'signer' for Azure::Storage::Default:Module`.